### 2024-02-27 #6421 @editorialbot
[PRE REVIEW]: GCIdentifier.jl: A Julia package for identifying molecular fragments from SMILES
**Submitting author:** @pw0908 (<a href="http://orcid.org/0000-0001-8628-6561">Pierre Walker</a>)
**Repository:** https://github.com/ClapeyronThermo/GCIdentifier.jl
**Branch with paper.md** (empty if default branch): main
**Version:** v0.2.0
