### 2024-03-08 #6453 @editorialbot
[REVIEW]: GCIdentifier.jl: A Julia package for identifying molecular fragments from SMILES
**Submitting author:** @pw0908 (<a href="http://orcid.org/0000-0001-8628-6561">Pierre Walker</a>)
**Repository:** https://github.com/ClapeyronThermo/GCIdentifier.jl
**Branch with paper.md** (empty if default branch): main
**Version:** v0.3.1
