### 2024-03-13 #6453 @pw0908
No clue what happened there but it looks like the pdf compiled just fine, the bot just didn't manage to link it to this issue: https://raw.githubusercontent.com/openjournals/joss-papers/joss.06453/joss.06453/10.21105.joss.06453.pdf
