### 2024-03-20 #6453 @mjohnson541
I was unable to install on a clean Julia 1.6.7 install from conda-forge. It looks like an issue with Cairo:
```
Incompatible library version: libpangocairo-1.0.0.dylib requires version 11603.0.0 or later, but libcairo.2.dylib provides version 2.0.0
```
