### 2024-03-20 #6453 @mjohnson541
As far as the documentation, on Home it claims "GCIdentifier.jl: An extensible Julia package for the modelling of fluids using thermodynamic equations of state." Which this software does not seem to do. I suspect this was accidentally copied from Clapeyron.jl. This should have a description of the software and what it can be used to do.
