### 2024-03-20 #6453 @pw0908
Hi Matt!

For the home page: that is a very silly mistake! We did copy our initital documentation template from Clapeyron, but it seems we forgot to change the opening lines. We have now fixed [this](https://github.com/ClapeyronThermo/GCIdentifier.jl/commit/f02d34c47d69039d096bf0833fe7eba9ec4ae8dd).

As for the installation error, are you certain of the Julia version? I've tried installation Julia 1.6.7 from conda-forge, but there is only 1.6.6, the next version up is 1.7.0. It's also strange that the tests fail on 1.6 as our automated [tests](https://github.com/ClapeyronThermo/GCIdentifier.jl/actions/runs/8354088073) include 1.6 and they've been passing. These only currently run on ubuntu so perhaps this is an issue with the OS?
