### 2024-03-20 ef3d601
run CI on mac an windows
