+++ .github/workflows/CI.yml
@@ -23,6 +23,8 @@ jobs:
           - 'nightly'
         os:
           - ubuntu-latest
+          - windows-latest
+          - macOS-latest
         arch:
           - x64
     steps:
