### 2024-03-20 #6453 @mjohnson541
Sorry, that was a mistake, I installed Julia v1.6.5 I am on osx though, which might be relevant.
