### 2024-03-20 #6453 @pw0908
We have now added macos and windows to our tests to see if this might be the issue. (MacOS just passed its tests)

I also made a mistake: 1.6.7 is the long-term support version of 1.6. But this was the version being tested on ubuntu either way.
