### 2024-03-20 #6453 @pw0908
@mjohnson541 Just tried installing GCIdentifier on a clean version of Julia 1.6.5 in Ubuntu without issue. It might just be a MacOS issue (are you using an M-series mac?).
