### 2024-03-28 #6453 @Arrondissement5etDemi
Hi Pierre,

For the software paper, could you please cite some packages with similar functionalities in other languages since they are mentioned in the paper?

Thanks!

Best regards,
Haina
