+++ paper/paper.md
@@ -41,7 +41,7 @@ The output from this function can then be used in other packages, such as Clapey
 
 
 # Statement of need
-Group-contribution approaches are vital when it comes to computer-aided molecular design (CAMD) of, for example, novel refrigerants [@sahinidisDesignAlternativeRefrigerants2003] or in drug discovery[@houADMEEvaluationDrug2004]. Here, the assignment of groups must be done thousands of times and, in some cases, for rather complex molecules. This is the primary motivator for the development of GCIdentifier. While other packages with similar functionalities have been developed in other languages, GCIdentifier stands apart for multiple reasons.
+Group-contribution approaches are vital when it comes to computer-aided molecular design (CAMD) of, for example, novel refrigerants [@sahinidisDesignAlternativeRefrigerants2003] or in drug discovery[@houADMEEvaluationDrug2004]. Here, the assignment of groups must be done thousands of times and, in some cases, for rather complex molecules. This is the primary motivator for the development of GCIdentifier. While other packages[@degenArtCompilingUsing2008;@liuBreakOrderBuild2017;@mullerFlexibleHeuristicAlgorithm2019] with similar functionalities have been developed in other languages, GCIdentifier stands apart for multiple reasons.
 
 GCIdentifier is the first of such packages to be compatible with multiple group-contribution approaches, such as UNIFAC and SAFT-$\gamma$ Mie. By standardising the representation of groups using SMARTS and leveraging the powerful MolecularGraph [@seiji_matsuoka_2024_10478701] package, our group-identification code can be used with any existing group-contribution thermodynamic model. This extends to group-contribution approaches which require information about the connectivity between groups [@sauerComparisonHomoHeterosegmented2014] where, by simply specifying `connectivity=true` within the `get_groups_from_name` function, the connectivity matrix between groups will automatically be generated.
 
