+++ paper/paper.bib
@@ -222,3 +222,60 @@ Section: New Results},
 	date = {2024-01},
 	doi = {10.5281/zenodo.10478701},
 }
+
+@article{mullerFlexibleHeuristicAlgorithm2019,
+	title = {Flexible heuristic algorithm for automatic molecule fragmentation: application to the {UNIFAC} group contribution model},
+	volume = {11},
+	issn = {1758-2946},
+	url = {https://doi.org/10.1186/s13321-019-0382-3},
+	doi = {10.1186/s13321-019-0382-3},
+	shorttitle = {Flexible heuristic algorithm for automatic molecule fragmentation},
+	abstract = {A priori calculation of thermophysical properties and predictive thermodynamic models can be very helpful for developing new industrial processes. Group contribution methods link the target property to contributions based on chemical groups or other molecular subunits of a given molecule. However, the fragmentation of the molecule into its subunits is usually done manually impeding the fast testing and development of new group contribution methods based on large databases of molecules. The aim of this work is to develop strategies to overcome the challenges that arise when attempting to fragment molecules automatically while keeping the definition of the groups as simple as possible. Furthermore, these strategies are implemented in two fragmentation algorithms. The first algorithm finds only one solution while the second algorithm finds all possible fragmentations. Both algorithms are tested to fragment a database of 20,000+ molecules for use with the group contribution model Universal Quasichemical Functional Group Activity Coefficients ({UNIFAC}). Comparison of the results with a reference database shows that both algorithms are capable of successfully fragmenting all the molecules automatically. Furthermore, when applying them on a larger database it is shown, that the newly developed algorithms are capable of fragmenting structures previously thought not possible to fragment.},
+	pages = {57},
+	number = {1},
+	journaltitle = {Journal of Cheminformatics},
+	shortjournal = {J Cheminform},
+	author = {Müller, Simon},
+	urldate = {2024-03-28},
+	date = {2019-08-20},
+	langid = {english},
+	keywords = {Cheminformatics, Group contribution method, Incrementation, Molecule fragmentation, Property prediction, {RDKit}, {UNIFAC}},
+}
+
+@article{degenArtCompilingUsing2008,
+	title = {On the Art of Compiling and Using 'Drug-Like' Chemical Fragment Spaces},
+	volume = {3},
+	rights = {Copyright © 2008 {WILEY}-{VCH} Verlag {GmbH} \& Co. {KGaA}, Weinheim},
+	issn = {1860-7187},
+	url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/cmdc.200800178},
+	doi = {10.1002/cmdc.200800178},
+	abstract = {To improve current methods for the decomposition of molecules into fragments, we compiled a new and more elaborate set of rules for the breaking of retrosynthetically interesting chemical substructures ({BRICS}). We also incorporated further medicinal chemistry concepts and compiled differently sized sets of diverse high-quality fragments. Relative to existing methods, {BRICS} performs much better in retrieving compounds from various large and diverse query sets.},
+	pages = {1503--1507},
+	number = {10},
+	journaltitle = {{ChemMedChem}},
+	author = {Degen, Jörg and Wegscheid-Gerlach, Christof and Zaliani, Andrea and Rarey, Matthias},
+	urldate = {2024-03-28},
+	date = {2008},
+	note = {\_eprint: https://onlinelibrary.wiley.com/doi/pdf/10.1002/cmdc.200800178},
+	keywords = {chemical fragment spaces, computer-aided molecular design, fragment-based lead discovery, retrosynthesis, shredding},
[…]
