+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"api/","page":"API","title":"API","text":"CurrentModule = GCIdentifier","category":"page"},{"location":"api/#Contents","page":"API","title":"Contents","text":"","category":"section"},{"location":"api/","page":"API","title":"API","text":"Pages = [\"api.md\"]","category":"page"},{"location":"api/#Index","page":"API","title":"Index","text":"","category":"section"},{"location":"api/","page":"API","title":"API","text":"Pages = [\"api.md\"]","category":"page"},{"location":"api/#types-and-methods","page":"API","title":"types and methods","text":"","category":"section"},{"location":"api/","page":"API","title":"API","text":"GCIdentifier.GCPair\nGCIdentifier.get_groups_from_smiles\nGCIdentifier.get_groups_from_name\nGCIdentifier.find_missing_groups_from_smiles\nGCIdentifier.get_grouplist\nGCIdentifier.@gcstring_str\nGCIdentifier.group_replace","category":"page"},{"location":"api/#GCIdentifier.GCPair","page":"API","title":"GCIdentifier.GCPair","text":"GCPair\n\nStruct used to hold a description of a group. contains the SMARTS string necessary to match the group within a SMILES query, and the assigned name.\n\n\n\n\n\n","category":"type"},{"location":"api/#GCIdentifier.get_groups_from_smiles","page":"API","title":"GCIdentifier.get_groups_from_smiles","text":"get_groups_from_smiles(smiles::String,groups;connectivity = false)\n\nGiven a SMILES string and a group list (groups::Vector{GCPair}), returns a list of groups and their corresponding amount.\n\nIf connectivity is true, then it will additionally return a vector containing the amount of bonds between each pair.\n\nExamples\n\njulia> get_groups_from_smiles(\"CCO\",UNIFACGroups)\n(\"CCO\", [\"CH3\" => 1, \"CH2\" => 1, \"OH(P)\" => 1])\n\njulia> get_groups_from_smiles(\"CCO\",JobackGroups,connectivity = true)\n(\"CCO\", [\"-CH3\" => 1, \"-CH2-\" => 1, \"-OH (alcohol)\" => 1], [(\"-CH3\", \"-CH2-\") => 1, (\"-CH2-\", \"-OH (alcohol)\") => 1])\n\n\n\n\n\n","category":"function"},{"location":"api/#GCIdentifier.get_groups_from_name","page":"API","title":"GCIdentifier.get_groups_from_name","text":"get_groups_from_name(name::String,groups;connectivity = false)\n\nGiven a molecule name and a group list (groups::Vector{GCPair}), returns a list of groups and their corresponding amount.\n\nIf connectivity is true, then it will additionally return a vector containing the amount of bonds between each pair.\n\nNote: Can only be used if ChemicalIdentifiers is also called.\n\nExamples\n\njulia> get_groups_from_name(\"ethanol\",UNIFACGroups)\n(\"ethanol\", [\"CH3\" => 1, \"CH2\" => 1, \"OH(P)\" => 1])\n\njulia> get_groups_from_name(\"ethanol\",JobackGroups,connectivity = true)\n(\"ethanol\", [\"-CH3\" => 1, \"-CH2-\" => 1, \"-OH (alcohol)\" => 1], [(\"-CH3\", \"-CH2-\") => 1, (\"-CH2-\", \"-OH (alcohol)\") => 1])\n\n\n\n\n\n","category":"function"},{"location":"api/#GCIdentifier.find_missing_groups_from_smiles","page":"API","title":"GCIdentifier.find_missing_groups_from_smiles","text":"find_missing_groups_from_smiles(smiles::String, gr
[…]
