+++ README.md
@@ -71,3 +71,4 @@ Pkg> rm GCIdentifier
 If you'd like to make a contribution to GCIdentifier, you can:
 * Report an issue: If you encounter an error in which groups are assigned or any other type of error, feel free to raise an issue and one of the developers will try to address it. If you've found the source of the error and fixed it yourself, feel free to also open a Pull Request so that it can be reviewed and pushed to the main branch.
 * Implement your own modifications: We are always open to making changes to the source code and documentation! If you'd like to add your own groups and/or functionalities to the package, feel free to open a pull request and one of the developers will review it before merging it with the main branch.
+* General support: If you have questions regarding the usage of this package or are having difficulties getting started, feel free to open a Discussion on GitHub or contact us directly on Zulip (link in badge above)!
