### 2024-03-28 #6453 @pw0908
Thank you both @Arrondissement5etDemi and @moyner for your comments! In the latests commits to GCIdentifier, we have addressed most of these issues.

For the comment relating to the tests, I haven't thought of a fix yet. They are occurring when we perform our general tests with UNIFAC where we just want to see, for a range of molecules, if the assignment is correct. We're not really testing for this warning at this stage. Do you think this warning is helpful @moyner? It is very common for there to be multiple combinations of groups to cover all atoms, it's just some of these use fewer groups total, and this is the one we pick. For example, in the case of acetone for UNIFAC, there is an acetone group, but there are also methyl and ketone groups which could also be used to represent acetone. We pick the acetone group since it represents the entire molecule with one group, but is it useful for the users to know that there is more than one possible combination? If not, then we'll suppress the warning.
