### 2024-03-29 #6453 @mjohnson541
Alright I was able to install it and test things properly. I think it looks good except for the summary in the paper which I have some detailed comments on:

```
“Obtaining chemical properties of species, such as heat capacities (Joback & Reid, 1987) or solvation free energies (Platts et al., 2000), will typically involve a set of parameters that represent a given species. Unfortunately, in these cases, the parameters obtained only apply to a specific species and cannot be transferred to others.”

The way this is stated is confusing:
1)	Because the references are group contribution schemes that are applicable to many different species.
2)	I don’t think it’s clear to a typical reader what “a set of parameters that represent a given species” is. If you’re referring to something like Abraham parameters, such a decomposition is not obvious to a typical reader. I think an illustrative example ideally on a property where the group decomposition makes a lot of intuitive sense would go a long way here.
3)	I feel like the essence of these sentences can be rephrased unfortunately as, “Obtaining chemical properties of a species only gives us chemical properties of that species.” To a general reader this leaves open a number of simple contextual questions that are a little silly for those of us that work in the area, but are in general valid questions for example: “Why don’t you just measure the chemical properties of the species you need?” “Why do you need these properties?”

It also seems funny to cite these papers, but not Benson’s scheme that they were based on: https://doi.org/10.1063/1.1744539

“An alternative approach would be to split the molecule into moieties, known as groups, each of which will have their own parameters associated with them. The combination of these groups (and their associated parameters) will represent the entire molecule. The benefit is that these groups can be combined in different ways such that they represent a new molecule”

An alternative to what approach?

The last sentence is true, but I think it will be unclear to a general audience because 1) the difference between the moieties and groups is not well established and 2) It’s not explicit that these groups aren’t only from the molecule in the first sentence. So a reader may think you are combining the groups on the original molecule in new ways instead of predicting a new molecule with the same groups.

“This type of approach is known as group contribution, where multiple examples of such approaches exist”

“Multiple” is technically correct, but I think “many” is much more appropriate.

```
