+++ paper/paper.bib
@@ -279,3 +279,20 @@ Section: New Results},
 	note = {Publisher: American Chemical Society},
 	file = {Full Text PDF:/Users/pierrewalker/Zotero/storage/Y7UTSAUG/Liu et al. - 2017 - Break Down in Order To Build Up Decomposing Small.pdf:application/pdf},
 }
+
+@article{bensonAdditivityRulesEstimation1958,
+	title = {Additivity Rules for the Estimation of Molecular Properties. Thermodynamic Properties},
+	volume = {29},
+	issn = {0021-9606},
+	url = {https://doi.org/10.1063/1.1744539},
+	doi = {10.1063/1.1744539},
+	abstract = {A general limiting law is proposed which can be used to systematize the various laws of additivity of molecular properties. This can be stated as follows: ``If φ is a molecular property, then for the disproportionation reaction: {RNR}+{SNS}⇄2RNS, Δφ→0 as the separation between R and S becomes large compared to their dimensions.It is shown that the zero‐order approximation is equivalent to the law of additivity of atomic properties, the first‐order approximation to the law of additivity of bond properties, the second‐order approximation to the law of additivity of group properties, and so forth. It is shown that for Cp and S° (ideal gases), the additivity of atomic properties works to about ±2 cal/mole‐°K, while the additivity of bond properties is usually good to about ±1 cal/mole‐°K. The latter also estimates ΔHf° to about ±3 kcal/mole. The group additivity relation is generally obeyed to within ±0.5 cal/mole‐°K for Cp and S° and about ±0.6 kcal/mole for ΔHf°. Tables are presented for each of the partial properties at 25°C.The agreements found for the various additivity rules is examined from a molecular point of view and certain extensions and limitations indicated. The application and utility of the rules in estimation of thermodynamic properties is discussed.The estimation of bond dissociation energies is possible with the additivity rules as are the thermodynamic properties of free radicals. An application of the rules to species {NR}2 and {NS}2 show that R and S may be ordered according to their general bond‐weakening or bond‐strengthening properties. For systems where N is C2H4 it is shown that the first and second bond dissociation energies for {RC}2H4R is very likely a constant = 59.5 kcal, the excitation energy of C2H4.},
+	pages = {546--572},
+	number = {3},
+	journaltitle = {The Journal of Chemical Physics},
+	shortjournal = {The Journal of Chemical Physics},
+	author = {Benson, Sidney W. and Buss, Jerry H.},
+	urldate = {2024-03-29},
+	date = {1958-09-01},
+	file = {Snapshot:/Users/pierrewalker/Zotero/storage/A4FR3R5S/Additivity-Rules-for-the-Estimation-of-Molecular.html:text/html},
+}
