### 2024-03-30 #6453 @mjohnson541
> Hi Matt, thanks a lot for your comments! I've now updated the summary based on your comments. The first paragraph was a bit hard to deal with as it is challenging bringing the general reader up to speed on what group-contribution is, how it's developed and why it's useful, without overwhelming them with equations. I decided to go with the example of Joback's method with a real molecule (ibuprofen). I'm not sure if I should add more detail as I think there might be a lot already. Let me know what you think!

Thanks!

I think is fine. More generally, you might also benefit from tweaking it to give the reader of the summary a better sense of the fact that group contribution schemes and this tool have more applications than just heat capacity and solvation prediction. Whether that is just changing the language or adding a few more examples from literature.
