+++ paper/paper.md
@@ -27,7 +27,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-GCIdentifier.jl is an open-source toolkit for the automatic identification of group fragments based on the name of a molecule or its SMILES. Obtaining chemical properties of species, such as heat capacities [@bensonAdditivityRulesEstimation1958] or solvation free energies [@plattsEstimationMolecularLinear2000], will typically involve a set of parameters that represent a given species. For example, ideal isobaric heat capacities over a range of temperature of a pure component can be obtained using Reid polynomials with just four parameters ($a$, $b$, $c$ and $d$). Unfortunately, in this case, the parameters obtained are only applicable to a specific species and cannot be transferred to others (i.e. the $a$, $b$, $c$, $d$ parameters for water cannot then be used to model ibuprofen). A solution to this would be to split a set of molecules with similar chemical structures into moieties, known as groups, each of which will have their own parameters associated with them and adjust these parameters against experimental data for all of these molecules. The combination of these groups (and their associated parameters) can then be used to predict the properties of ibuprofen. In the case of the Joback method[@jobackEstimationPureComponentProperties1987], the Reid polynomial parameters can be obtained by summing over the group-specific parameters ($a_i$, $b_i$, $c_i$ and $d_i$) weighted by the occurrence of those groups in a species. The benefit of such approaches is that these groups can be combined many different ways such that they represent a larger variety of molecules. This type of approach is known as group contribution, where many examples of such approaches exist [@weidlichModifiedUNIFACModel1987;@walkerNewPredictiveGroupContribution2020;@chungGroupContributionMachine2022;@papaioannouGroupContributionMethodology2014]. An example of this process is shown in figure 1.
+GCIdentifier.jl is an open-source toolkit for the automatic identification of group fragments based on the name of a molecule or its SMILES. Obtaining chemical properties of species, such as heat capacities [@bensonAdditivityRulesEstimation1958] or solvation free energies [@plattsEstimationMolecularLinear2000], will typically involve a set of parameters that represent a given species. For example, ideal isobaric heat capacities over a range of temperature of a pure component can be obtained using Reid polynomials with just four parameters ($a$, $b$, $c$ and $d$). Unfortunately, in this case, the parameters obtained are only applicable to a specific species and cannot be transferred to others (i.e. the $a$, $b$, $c$, $d$ parameters for water cannot then be used to model ibuprofen). A solution to this would be to split a set of molecules with similar chemical structures into moieties, known as groups, each of which will have their own parameters associated with them and adjust these parameters against experimental data for all of these molecules. The combination of these groups (and their associated parameters) can then be used to predict the properties of ibuprofen. In the case of the Joback method[@jobackEstimationPureComponentProperties1987], the Reid polynomial parameters can be obtained by summing over the group-specific parameters ($a_i$, $b_i$, $c_i$ and $d_i$) weighted by the occurrence of those groups in a species. The benefit of such approaches is that these groups can be combined many different ways such that they represent a larger variety of molecules. This type of approach is known as group contribution, where many examples of such approaches exist [@weidlichModifiedUNIFACModel1987;@walkerNewPredictiveGroupContribution2020;@chungGroupContributionMachine2022;@papaioannouGroupContributionMethodology2014] which can be used to predict a range of properties such as pharmaceutical solubilities[@wehbePhaseBehaviourPHsolubility2022], interfacial tensions[@rehnerSurfactantModelingUsing2021] and thermal conductivites[@hoppThermalConductivityEntropy2019]. An example of this process is shown in figure 1.
 
 ![Fragmentation of ibuprofen into UNIFAC groups.](figures/ibuprofen.pdf)
 
