+++ paper/paper.bib
@@ -296,3 +296,55 @@ Section: New Results},
 	date = {1958-09-01},
 	file = {Snapshot:/Users/pierrewalker/Zotero/storage/A4FR3R5S/Additivity-Rules-for-the-Estimation-of-Molecular.html:text/html},
 }
+@article{hoppThermalConductivityEntropy2019,
+	title = {Thermal Conductivity from Entropy Scaling: A Group-Contribution Method},
+	volume = {58},
+	issn = {0888-5885},
+	url = {https://doi.org/10.1021/acs.iecr.9b04289},
+	doi = {10.1021/acs.iecr.9b04289},
+	shorttitle = {Thermal Conductivity from Entropy Scaling},
+	abstract = {Entropy scaling has proven to be a powerful method for calculating transport properties. The applicability of the entropy scaling approach to predict the viscosity, thermal conductivity, and self-diffusion coefficients of pure substances based on substance-specific parameters over the last years was convincingly demonstrated in the literature. In this work, we derive a predictive method for the thermal conductivity based on entropy scaling. The model is developed as a group-contribution approach where substances are considered to be composed of chemical (functional) groups. The excess entropy is calculated using the group-contribution {PCP}-{SAFT} equation of state. The model is applicable for gaseous phases and for liquid-phase conditions covering wide ranges of temperature and pressure. We consider pure fluids from various chemical families, namely, alkanes, branched alkanes, cyclic alkanes, alkenes, aldehydes, aromatics, esters, ethers, ketones and alcohols, and some individual substances, such as water, carbon dioxide, and the like. We propose parameters of 29 chemical groups by considering 231 substances with more than 50,000 experimental data points. The group-contribution method for the thermal conductivity proposed in this work is shown to be in convincing agreement with experimental data with 6.17\% average absolute deviation for all considered data points.},
+	pages = {20441--20449},
+	number = {44},
+	journaltitle = {Industrial \& Engineering Chemistry Research},
+	shortjournal = {Ind. Eng. Chem. Res.},
+	author = {Hopp, Madlen and Gross, Joachim},
+	urldate = {2024-03-30},
+	date = {2019-11-06},
+	note = {Publisher: American Chemical Society},
+}
+
+@article{rehnerSurfactantModelingUsing2021,
+	title = {Surfactant Modeling Using Classical Density Functional Theory and a Group Contribution {PC}-{SAFT} Approach},
+	volume = {60},
+	issn = {0888-5885},
+	url = {https://doi.org/10.1021/acs.iecr.1c00169},
+	doi = {10.1021/acs.iecr.1c00169},
[…]
