### 2024-04-01 #6453 @zhubonan
Hi @pw0908 there are a few more steps before I can move the submission to the next stage. Please make a release of the package and deposit to a publicly available repository (for example Zenodo) to get an DOI. This DOI will be linked to the JOSS publication. Please see the checklist and make sure the metadata matches with that of the paper.

There is one thing I noticed when proofreading the manuscript. The package name is initially referred as `GCIdentifier.jl` but `GCIdentifier` is used subsequently. I think it is better to choose one of the two.
