### 2024-04-03 #6453 @zhubonan
Hi @pw0908 Could you please edit the metadata of the Zenodo entry such that it matches that of the paper.
In particular, the title, the author list (please also add their ORCIDs) should match the paper, and the licence should match your actual software licencee.

Note that there is no need to make a new release (since the files remain unchanged). You can check recently published JOSS paper as examples, such as this one: https://joss.theoj.org/papers/10.21105/joss.06294
