### 2024-04-03 #6453 @pw0908
Do you want us to make a manual zenodo? This is the automatic one generated by GitHub which I can't modify unless I make a new release.
