### 2024-04-03 3ee9020
Add zenodo json and bump
