### 2024-04-03 #6453 @pw0908
@zhubonan It's all been fixed now although the version and zenodo needed to be updated:
version: v0.3.1
zenodo: 10.5281/zenodo.10914015

I have also updated the ORCIDs although one of our authors doesn't have an ORCID. I've seen a few JOSS papers where not all authors had an ORCID so I assume this is ok?
