### 2024-04-04 #6453 @Kevin-Mattheus-Moerman
as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I have checked this review, your repository, the archive link, and the paper. Most seems in order and I only have below point that requires your attention:

- [ ] It looks like `conductivites` should be `conductivities` please check
