### 2024-04-04 #6453 @pw0908
Yes! Thank you for catching this. I have now fixed the typo.
