### 2024-04-05 #6453 @Kevin-Mattheus-Moerman
One more thing @pw0908 can you check that you always use spaces after a reference? See below I highlight some correct and incorrect locations:
![temp](https://github.com/openjournals/joss-reviews/assets/8392709/ea98bf9d-644d-4818-a9fe-0e949c066c03)
